refactor attribute parsing to its own module
